requirements.txt
asgiref==3.4.1
dj-rest-auth==2.1.11
Django==3.2.6
django-cors-headers==3.8.0
django-sslserver==0.22
djangorestframework==3.12.4
ecdsa==0.17.0
jose==1.0.0
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.1
pyasn1==0.4.8
pydantic==1.8.2
python-jose==3.3.0
pytz==2021.1
qrcode==7.3
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0